v1.005
Нормально работает

+++ Calc/WebCalc/Models/CalcViewModel.cs
@@ -8,9 +8,9 @@ namespace WebCalc.Models
     {
         public string Result { get; set; }
 
-        public int X { get; set; }
+        public double X { get; set; }
 
-        public int Y { get; set; }
+        public double Y { get; set; }
 
         public string Operation { get; set; }
